feat: add UWP application picker stub
Adds a "Add UWP Application" button to the main window and connects it to a new callback. A stub `show_uwp_picker` function is implemented for Windows, which currently displays an error message indicating that the feature is not yet complete. A non-Windows stub is also provided.

This lays the groundwork for UWP application integration and facilitates future development in tasks 10.2 and 10.3.